requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
black==21.12b0
cfgv==3.3.1
click==8.1.3
colorama==0.4.4
coverage==6.2
distlib==0.3.4
filelock==3.4.2
flake8==4.0.1
freezegun==1.1.0
identify==2.4.3
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.16.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pynput==1.7.6
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-freezegun==0.4.2
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
virtualenv==20.13.0
websocket-client==1.3.3
websockets==10.3
